field types to Int; simpligy GisFeature converter

+++ DUI3-DX/Converters/ArcGIS/Speckle.Converters.ArcGIS3/Layers/TableToSpeckleConverter.cs
@@ -43,7 +43,7 @@ public class StandaloneTableToSpeckleConverter
       if (field.IsVisible)
       {
         string name = field.Name;
-        attributes[name] = field.Type;
+        attributes[name] = (int)field.Type;
       }
     }
